plop
git-svn-id: http://svn.net-core.org/repos/t-engine4@2259 51575b47-30f0-44d4-a5cc-537603b46e54

+++ src/map.c
@@ -581,7 +581,7 @@ static int map_set_grid(lua_State *L)
 		// We use the pointer value directly as an index
 		if (map->grids[x][y][i])
 		{
-			lua_pushnumber(L, (long)map->grids[x][y][i]);
+			lua_pushnumber(L, (long long)map->grids[x][y][i]);
 			lua_pushnil(L);
 			lua_settable(L, 6); // Access the list of all mos for the map
 		}
